Assignment/requirements/base.txt
django==2.2.1
django-environ==0.4.5
psycopg2-binary
celery==4.3.0
django-celery-results==1.0.1
djangorestframework
